Refactor Tree Layout 1
