+++ packages/layout/src/tree-layout/layout-component/dummy-component.js
