--- packages/layout/src/tree-layout/src/index.html
@@ -1,44 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-
-<head>
-	<meta charset="UTF-8">
-	<meta name="viewport" content="width=device-width, initial-scale=1.0">
-	<meta http-equiv="X-UA-Compatible" content="ie=edge">
-	<title>Layout</title>
-	<style>
-		html,body{
-            height: 100%;
-            width: 100%;
-            background-color: #fff;
-        }
-        .flex-container {
-            position: relative;
-            width: 100%;
-            height: 100%;
-            padding: 0;
-            margin: 0;
-        }
-        
-        .Aligner {
-            display: flex;
-            align-items: center;
-            justify-content: center;
-        }
-        
-       #board{
-           background: snow;
-           border: 1px solid #ccc;
-       }
-	</style>
-</head>
-
-<body>
-    <div class="flex-container Aligner">
-        <div id='board' class="Aligner-item"></div>
-    </div>
-	<script src="./layout.js"></script>
-	<script src="./test.1.js"></script>
-</body>
-
-</html>
